Refactor: hoist API auth-context helpers into ApiControllerBase (DRY)
The IsApiKeyAuthenticated / GetApiKeyName / GetCurrentApiKeyAsync /
GetCurrentUserAsync helpers were copy-pasted across SynchronisationController,
HistoryController, MetaverseController and ServiceSettingsController. Consolidate
them into a single abstract ApiControllerBase that the four controllers inherit.

- New ApiControllerBase(JimApplication application, ILogger logger) : ControllerBase
  exposing protected Application/Logger and the four helpers (one definition).
- The four controllers now derive from ApiControllerBase and pass their injected
  application + typed logger to it; their own _application/_logger fields are kept
  for their (extensive) body usage, so no body churn. Behaviour is identical.
- Removed the duplicated helper bodies and the resulting unused usings
  (System.Security.Claims, JIM.Models.Security) from the three controllers that no
  longer reference them.

No functional change. Build clean (0 warnings); Web.Api 751 and Worker 1645 green.
