+++ src/JIM.Web/Controllers/Api/SynchronisationController.cs
@@ -40,7 +40,7 @@ public class SynchronisationController(
     ILogger<SynchronisationController> logger,
     JimApplication application,
     IExpressionEvaluator expressionEvaluator,
-    ICredentialProtectionService credentialProtection) : ControllerBase
+    ICredentialProtectionService credentialProtection) : ApiControllerBase(application, logger)
 {
     private readonly ILogger<SynchronisationController> _logger = logger;
     private readonly JimApplication _application = application;
@@ -3270,96 +3270,6 @@ public class SynchronisationController(
     #endregion
 
     #region Private Helpers
-
-    /// <summary>
-    /// Checks if the current authentication is via API key.
-    /// </summary>
-    private bool IsApiKeyAuthenticated()
-    {
-        return User.HasClaim("auth_method", "api_key");
-    }
-
-    /// <summary>
-    /// Gets the API key name if authenticated via API key.
-    /// </summary>
-    private string? GetApiKeyName()
-    {
-        if (!IsApiKeyAuthenticated())
-            return null;
-
-        return User.Identity?.Name;
-    }
-
-    /// <summary>
-    /// Gets the current API key entity if authenticated via API key.
-    /// </summary>
-    private async Task<JIM.Models.Security.ApiKey?> GetCurrentApiKeyAsync()
-    {
-        if (!IsApiKeyAuthenticated())
-            return null;
-
-        // The API key ID is stored in the NameIdentifier claim
-        var apiKeyIdClaim = User.FindFirst(System.Security.Claims.ClaimTypes.NameIdentifier)?.Value;
-        if (string.IsNullOrEmpty(apiKeyIdClaim) || !Guid.TryParse(apiKeyIdClaim, out var apiKeyId))
-            return null;
-
-        return await _application.Repository.ApiKeys.GetByIdAsync(apiKeyId);
-    }
-
-    /// <summary>
-    /// Resolves the current user from JWT claims by looking up their SSO identifier in the Metaverse.
-    /// Returns null for API key authentication (which is valid - use IsApiKeyAuthenticated() to check).
-    /// </summary>
-    private async Task<JIM.Models.Core.MetaverseObject?> GetCurrentUserAsync()
-    {
-        if (User.Identity?.IsAuthenticated != true)
-            return null;
-
-        // API key authentication doesn't map to a Metaverse user object
-        // This is valid - the caller should check IsApiKeyAuthenticated() separately
-        if (IsApiKeyAuthenticated())
-        {
-            _logger.LogDebug("API key authentication detected - no Metaverse user lookup needed");
-            return null;
-        }
-
-        // Get the service settings to know which claim type contains the unique identifier
-        var serviceSettings = await _application.ServiceSettings.GetServiceSettingsAsync();
-        if (serviceSettings?.SSOUniqueIdentifierClaimType == null ||
-            serviceSettings.SSOUniqueIdentifierMetaverseAttribute == null)
-        {
-            _logger.LogError("Service settings are not configured for SSO claim mapping");
-            return null;
-        }
-
-        // Get the unique identifier from the JWT claims
-        var uniqueIdClaimValue = IdentityUtilities.GetSsoUniqueIdentifier(
-            User,
-            serviceSettings.SSOUniqueIdentifierClaimType);
-
-        if (string.IsNullOrEmpty(uniqueIdClaimValue))
-        {
-            _logger.LogWarning("JWT does not contain the expected claim: {ClaimType}",
-                serviceSettings.SSOUniqueIdentifierClaimType);
-            return null;
-        }
-
-        // Look up the user in the Metaverse
-        var userType = await _application.Metaverse.GetMetaverseObjectTypeAsync(
-            JIM.Models.Core.Constants.BuiltInObjectTypes.User,
-            false);
-
-        if (userType == null)
-        {
-            _logger.LogError("Could not find User object type in Metaverse");
-            return null;
-        }
-
-        return await _application.Metaverse.GetMetaverseObjectByTypeAndAttributeAsync(
-            userType,
-            serviceSettings.SSOUniqueIdentifierMetaverseAttribute,
-            uniqueIdClaimValue);
-    }
 
     /// <summary>
     /// Checks if a Container belongs to a Connected System, traversing the parent Container chain if necessary.
